fix bugs with eventlistener in modalwindow
