+++ src/index.js
@@ -8,14 +8,11 @@ import './js/team-list';
 import { getMovies, renderMoviesList } from './js/movies-list';
 import { searchMovie } from './js/searchMovie';
 import { renderLoadingSpinner } from './js/loadingSpinner';
-import { addToList } from './js/addToList';
 import { openModal, closeModal } from './js/details';
 
 // VARIABLES
 const searchForm = document.getElementById('search-form');
 export const moviesContainer = document.querySelector('.covers-container');
-
-const closeBtn = document.getElementById('close-modal');
 
 // STATE
 export const state = {
@@ -89,18 +86,3 @@ if (searchForm) {
 
 // Open modal
 moviesContainer.addEventListener('click', openModal);
-
-// Close modal
-closeBtn.addEventListener('click', closeModal);
-
-window.addEventListener('keydown', (event) => {
-  if (event.key === 'Escape') {
-    closeModal();
-  }
-});
-
-modal.addEventListener('click', (event) => {
-  if (event.target === modal) {
-    closeModal();
-  }
-});
