+++ src/partials/details.html
@@ -29,13 +29,17 @@
         <p id="movieAbout" class="details__about-content"></p>
       </div>
       <div class="details__btn-wrapper">
-        <button type="button" class="details__btn" id="btn-addToWatch">
-          add to<br />
-          Watched
-        </button>
-        <button type="button" class="details__btn" id="btn-addToQueue">
-          add to queue
-        </button>
+        <div>
+          <button type="button" class="details__btn" id="btn-addToWatch">
+            add to<br />
+            Watched
+          </button>
+        </div>
+        <div>
+          <button type="button" class="details__btn" id="btn-addToQueue">
+            add to queue
+          </button>
+        </div>
       </div>
     </div>
   </div>
